adapt the script for new api
