+++ pessimistic_proof/src/lib.rs
@@ -2,13 +2,15 @@ pub mod keccak;
 pub mod local_exit_tree;
 
 mod proof;
-pub use proof::{generate_full_proof_with_state, FullProofOutput, ProofError, State};
+pub use proof::{generate_full_proof_with_state, FullProofOutput, ProofError};
 
 pub mod test_utils;
+
+pub mod certificate;
+pub mod local_balance_tree;
 
 mod withdrawal;
 pub use withdrawal::{NetworkId, TokenInfo, Withdrawal};
 
-pub mod certificate;
-
-pub mod local_balance_tree;
+pub mod global_state;
+pub use global_state::State;
